+++ suomen-vaylat-app/src/components/layout/Header.jsx
@@ -17,13 +17,15 @@ import { ReactComponent as VaylaLogoEn } from './images/vayla_sivussa_en_white.s
 import { ReactComponent as VaylaLogoFi } from './images/vayla_sivussa_fi_white.svg';
 import { ReactComponent as VaylaLogoSv } from './images/vayla_sivussa_sv_white.svg';
 import { updateLayers } from '../../utils/rpcUtil';
+import { createBrowserHistory } from 'history';
+const history = createBrowserHistory();
 
 const StyledHeaderContainer = styled.div`
     height: 64px;
     display: grid;
     position: relative;
     grid-template-columns: 1fr 1fr 1fr;
-    box-shadow: rgb(0 0 0 / 16%) 0px 3px 6px, rgb(0 0 0 / 23%) 0px 3px 6px;
+    box-shadow: 0px 2px 4px #0000004D;
     @media ${props => props.theme.device.tablet} {
         height: 56px;
     };
@@ -37,19 +39,11 @@ const StyledHeaderButton = styled.div`
     display: flex;
     justify-content: center;
     align-items: center;
-    margin-right: 8px;
     background-color: transparent;
     border-radius: 50%;
     svg {
         color: ${props => props.theme.colors.mainWhite};
         font-size: 22px;
-    };
-    @media ${props => props.theme.device.mobileL} {
-        width: 40px;
-        height: 40px;
-        svg {
-            font-size: 22px;
-        };
     };
 `;
 
@@ -76,14 +70,11 @@ const StyledHeaderLogoContainer = styled.div`
     display: flex;
     justify-content: center;
     align-items: center;
+    a {
+        height: inherit;
+    };
     svg {
         height: inherit;
-        max-height: 60px;
-    };
-    @media ${props => props.theme.device.mobileL} {
-        svg {
-            max-height: 48px;
-        };
     };
 `;
 
@@ -100,6 +91,7 @@ const StyledRightCornerButtons = styled.div`
 
 const DesktopButtons = styled.div`
     display: flex;
+    gap: 4px;
     @media ${props => props.theme.device.mobileL} {
         display: none;
     };
@@ -162,9 +154,13 @@ export const Header = () => {
     };
 
     const setToMainScreen = () => {
+        let routerPrefix = '/';
+        if (process.env.REACT_APP_ROUTER_PREFIX) {
+            routerPrefix = process.env.REACT_APP_ROUTER_PREFIX;
+        }
         // remove all selected layers
         selectedLayers.forEach((layer) => {
-            channel.postRequest('MapModulePlugin.MapLayerVisibilityRequest', [layer.id, false]);
+            channel && channel.postRequest('MapModulePlugin.MapLayerVisibilityRequest', [layer.id, false]);
         });
 
         // set map center
@@ -176,13 +172,17 @@ export const Header = () => {
 
         // add start layers
         startState.selectedLayers.forEach((layerId) => {
-            channel.postRequest('MapModulePlugin.MapLayerVisibilityRequest', [layerId, true]);
+            channel && channel.postRequest('MapModulePlugin.MapLayerVisibilityRequest', [layerId, true]);
         });
 
         store.dispatch(setIsMainScreen());
+        history.push(routerPrefix);
         handleSelectGroup(null, lastSelectedTheme);
 
-        updateLayers(store, channel)
+        updateLayers(store, channel);
+
+        // Remove all features from map
+        channel && channel.postRequest('MapModulePlugin.RemoveFeaturesFromMapRequest', []);
     };
 
     return (
@@ -195,12 +195,23 @@ export const Header = () => {
                     <span>{strings.tooltips.showUserGuide}</span>
                 </ReactTooltip>
                 <StyledHeaderTitleContainer onClick={() => setToMainScreen()}>
-                        {strings.title}
+                    {strings.title}
                 </StyledHeaderTitleContainer>
                 <StyledHeaderLogoContainer>
-                    {   lang.current === 'fi' ? <VaylaLogoFi /> :
-                        lang.current === 'en' ? <VaylaLogoEn /> :
-                        lang.current === 'sv' ? <VaylaLogoSv /> : <VaylaLogoFi />}
+                    <a
+                        href={
+                            lang.current === 'fi' ? "https://vayla.fi/etusivu" :
+                            lang.current === 'en' ? "https://vayla.fi/en/frontpage" :
+                            lang.current === 'sv' ? "https://vayla.fi/sv/framsida" : "https://vayla.fi/etusivu"
+                        }
+                        target="_blank"
+                        rel="noreferrer">
+                        {
+                            lang.current === 'fi' ? <VaylaLogoFi /> :
+                            lang.current === 'en' ? <VaylaLogoEn /> :
+                            lang.current === 'sv' ? <VaylaLogoSv /> : <VaylaLogoFi />
+                        }
+                    </a>
                 </StyledHeaderLogoContainer>
                 <StyledRightCornerButtons>
                     <MobileButtons>
@@ -229,7 +240,7 @@ export const Header = () => {
                     {
                         isSubNavOpen &&
                         <StyledMobileNavContainer
-                            initial={{ y: -100, filter: "blur(1px)", opacity: 0 }}
+                            initial={{ y: -100, filter: "blur(10px)", opacity: 0 }}
                             animate={{ y: 0, filter: "blur(0px)", opacity: 1 }}
                             exit={{ y: -100, filter: "blur(10px)", opacity: 0 }}
                             transition={{