+++ suomen-vaylat-app/src/components/search/VKMSearchStyles.jsx
@@ -1,21 +1,32 @@
 import strings from '../../translations';
 
 export const VKMGeoJsonStyles = {
-    tie: {
-        stroke: {
-            width: 12,
-            color: '#910aa3'
+    road: {
+        tie: {
+            stroke: {
+                width: 12,
+                color: '#910aa3'
+            }
+        },
+        osa: {
+            stroke: {
+                width: 8,
+                color: '#ffc300'
+            }
+        },
+        etaisyys: {
+            image: {
+                radius: 14,
+                fill: {
+                    color: '#c73f00'
+                }
+            }
         }
     },
-    osa: {
-        stroke: {
-            width: 8,
-            color: '#ffc300'
-        }
-    },
-    etaisyys: {
+    track: {
         image: {
-            radius: 14,
+            shape: 5,
+            size: 3,
             fill: {
                 color: '#c73f00'
             }
@@ -24,32 +35,46 @@ export const VKMGeoJsonStyles = {
 };
 
 export const VKMGeoJsonHoverStyles = {
-    tie: {
-        inherit: true,
-        effect: 'darken',
-        content: [
-            { key: strings.search.vkm.tie, valueProperty: 'tie' },
-            { key: strings.search.vkm.osa, valueProperty: 'osa' },
-            { key: strings.search.vkm.ajorata, valueProperty: 'ajorata' }
-        ]
+    road: {
+        tie: {
+            inherit: true,
+            effect: 'darken',
+            content: [
+                { key: strings.search.vkm.tie, valueProperty: 'tie' },
+                { key: strings.search.vkm.osa, valueProperty: 'osa' },
+                { key: strings.search.vkm.ajorata, valueProperty: 'ajorata' }
+            ]
+        },
+        osa: {
+            inherit: true,
+            effect: 'darken',
+            content: [
+                { key: strings.search.vkm.tie, valueProperty: 'tie' },
+                { key: strings.search.vkm.osa, valueProperty: 'osa' },
+                { key: strings.search.vkm.ajorata, valueProperty: 'ajorata' }
+            ]
+        },
+        etaisyys: {
+            inherit: true,
+            effect: 'darken',
+            content: [
+                { key: strings.search.vkm.tie, valueProperty: 'tie' },
+                { key: strings.search.vkm.osa, valueProperty: 'osa' },
+                { key: strings.search.vkm.ajorata, valueProperty: 'ajorata' },
+                { key: strings.search.vkm.etaisyys, valueProperty: 'etaisyys' }
+            ]
+        }
     },
-    osa: {
+    track: {
         inherit: true,
         effect: 'darken',
         content: [
-            { key: strings.search.vkm.tie, valueProperty: 'tie' },
-            { key: strings.search.vkm.osa, valueProperty: 'osa' },
-            { key: strings.search.vkm.ajorata, valueProperty: 'ajorata' }
-        ]
-    },
-    etaisyys: {
-        inherit: true,
-        effect: 'darken',
-        content: [
-            { key: strings.search.vkm.tie, valueProperty: 'tie' },
-            { key: strings.search.vkm.osa, valueProperty: 'osa' },
-            { key: strings.search.vkm.ajorata, valueProperty: 'ajorata' },
-            { key: strings.search.vkm.etaisyys, valueProperty: 'etaisyys' }
+            { key: strings.search.vkm.sijaintiraide, valueProperty: 'sijaintiraide' },
+            { key: strings.search.vkm.kuvaus, valueProperty: 'sijaintiraide_kuvaus' },
+            { key: strings.search.vkm.tyyppi, valueProperty: 'sijaintiraide_tyyppi' },
+            { key: strings.search.vkm.ratanumero, valueProperty: 'ratanumero' },
+            { key: strings.search.vkm.ratakilometri, valueProperty: 'ratakilometri' },
+            { key: strings.search.vkm.ratametri, valueProperty: 'ratametri' },
         ]
     }
 };