+++ ReleaseNotes.md
@@ -1,8 +1,16 @@
 # Release Notes
 
+## 1.0.0
+
+For a full list of changes see: https://github.com/finnishtransportagency/suomen-vaylat/milestone/3?closed=1
+
+* Improve search component
+* Open legend when theme maps is selected
+* Improve share descriptions
+
 ## 0.3.0
 
-For a full lis of changes see: https://github.com/finnishtransportagency/suomen-vaylat/milestone/2?closed=1
+For a full list of changes see: https://github.com/finnishtransportagency/suomen-vaylat/milestone/2?closed=1
 
 * React-modal changed to own modal component
 * Fixed an issue for theme sharing